+++ cmd/ops/ops_test.go
@@ -1,17 +1,19 @@
-package plural_test
+package ops_test
 
 import (
 	"os"
 	"testing"
 
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/mock"
-	"gopkg.in/yaml.v2"
+	"github.com/pluralsh/plural-cli/pkg/common"
 
 	"github.com/pluralsh/plural-cli/cmd/plural"
+	clientcmd "github.com/pluralsh/plural-cli/pkg/client"
 	"github.com/pluralsh/plural-cli/pkg/manifest"
 	"github.com/pluralsh/plural-cli/pkg/test/mocks"
 	"github.com/pluralsh/plural-cli/pkg/utils/git"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
+	"gopkg.in/yaml.v2"
 
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -55,10 +57,15 @@ func TestListNodes(t *testing.T) {
 			client := mocks.NewClient(t)
 			kube := mocks.NewKube(t)
 			kube.On("Nodes").Return(test.nodes, nil)
-			app := plural.CreateNewApp(&plural.Plural{Client: client, Kube: kube})
+			app := plural.CreateNewApp(&plural.Plural{
+				Plural: clientcmd.Plural{
+					Client: client,
+					Kube:   kube,
+				},
+			})
 			app.HelpName = plural.ApplicationName
 			os.Args = test.args
-			res, err := captureStdout(app, os.Args)
+			res, err := common.CaptureStdout(app, os.Args)
 			assert.NoError(t, err)
 
 			assert.Equal(t, test.expectedResponse, res)
@@ -116,10 +123,15 @@ func TestTerminate(t *testing.T) {
 			client := mocks.NewClient(t)
 			kube := mocks.NewKube(t)
 			kube.On("Node", mock.AnythingOfType("string")).Return(test.node, nil)
-			app := plural.CreateNewApp(&plural.Plural{Client: client, Kube: kube})
+			app := plural.CreateNewApp(&plural.Plural{
+				Plural: clientcmd.Plural{
+					Client: client,
+					Kube:   kube,
+				},
+			})
 			app.HelpName = plural.ApplicationName
 			os.Args = test.args
-			res, err := captureStdout(app, os.Args)
+			res, err := common.CaptureStdout(app, os.Args)
 			assert.NoError(t, err)
 
 			assert.Equal(t, test.expectedResponse, res)
