+++ cmd/proxy/proxy_test.go
@@ -1,8 +1,11 @@
-package plural_test
+package proxy_test
 
 import (
+	clientcmd "github.com/pluralsh/plural-cli/pkg/client"
 	"os"
 	"testing"
+
+	"github.com/pluralsh/plural-cli/pkg/common"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
@@ -51,10 +54,15 @@ func TestProxyList(t *testing.T) {
 			client := mocks.NewClient(t)
 			kube := mocks.NewKube(t)
 			kube.On("ProxyList", mock.AnythingOfType("string"), mock.AnythingOfType("string")).Return(test.proxyList, nil)
-			app := plural.CreateNewApp(&plural.Plural{Client: client, Kube: kube})
+			app := plural.CreateNewApp(&plural.Plural{
+				Plural: clientcmd.Plural{
+					Client: client,
+					Kube:   kube,
+				},
+			})
 			app.HelpName = plural.ApplicationName
 			os.Args = test.args
-			res, err := captureStdout(app, os.Args)
+			res, err := common.CaptureStdout(app, os.Args)
 			assert.NoError(t, err)
 
 			assert.Equal(t, test.expectedResponse, res)
